Issue #1: Renamed modules.
