+++ package.json
@@ -92,6 +92,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.23.0",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
     "rimraf": "^5.0.5",
