+++ src/pages/AppPage/AppPage.tsx
@@ -1,16 +1,16 @@
 import { AppWrapper } from 'src/ui/wrappers/AppWrapper';
 import { FullScreenPageLayout } from 'src/ui/layouts/FullScreenPageLayout';
-import { AppRouterWrapper } from 'src/ui/wrappers/AppRouterWrapper';
+import { AppRouterNavigator } from 'src/ui/wrappers/AppRouterNavigator';
 import { ShowData } from 'src/components/Data/ShowData';
 
 export function AppPage() {
   return (
     <AppWrapper>
       <FullScreenPageLayout>
-        <AppRouterWrapper>
+        <AppRouterNavigator>
           {/* Placeholder for the main app page content */}
           <ShowData />
-        </AppRouterWrapper>
+        </AppRouterNavigator>
       </FullScreenPageLayout>
     </AppWrapper>
   );
