+++ src/ui/wrappers/AppRouterNavigator/AppRouterNavigator.module.scss
