+++ src/ui/wrappers/AppRouterNavigator/AppRouterNavigatorWaiter/AppRouterNavigatorWaiter.tsx
@@ -7,14 +7,14 @@ import { TPropsWithClassName } from 'src/core/types';
 import { LoaderSplash } from 'src/ui/Basic';
 import { useAppSessionStore } from 'src/store';
 
-type TAppRouterWrapperWaiterProps = TPropsWithClassName;
+type TAppRouterNavigatorWaiterProps = TPropsWithClassName;
 
-export const AppRouterWrapperWaiter: React.FC<TAppRouterWrapperWaiterProps> = observer((props) => {
+export const AppRouterNavigatorWaiter: React.FC<TAppRouterNavigatorWaiterProps> = observer((props) => {
   const { className } = props;
   const appSessionStore = useAppSessionStore();
   const { themeMode } = appSessionStore;
   return (
-    <Box className={classNames(className, 'AppRouterWrapperWaiter')}>
+    <Box className={classNames(className, 'AppRouterNavigatorWaiter')}>
       <LoaderSplash show mode="cover" fullSize themeMode={themeMode} />
     </Box>
   );
