+++ src/ui/wrappers/AppRouterNavigator/AppRouterNavigatorWaiter/index.ts
@@ -1,0 +1,1 @@
+export * from './AppRouterNavigatorWaiter';
