+++ src/ui/wrappers/AppRouterNavigator/index.ts
@@ -1,0 +1,1 @@
+export * from './AppRouterNavigator';
