--- src/ui/wrappers/AppRouterWrapper/AppRouterWrapperWaiter/index.ts
@@ -1,1 +1,0 @@
-export * from './AppRouterWrapperWaiter';
