--- src/ui/wrappers/AppRouterWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './AppRouterWrapper';
