+++ src/ui/wrappers/CoreUiWrapper/CoreUiWrapper.tsx
@@ -3,15 +3,15 @@ import { CssBaseline, Typography } from '@mui/material';
 import classNames from 'classnames';
 
 import { WithToastsWrapper } from 'src/ui/Basic';
-import styles from './CoreWrapper.module.scss';
+import styles from './CoreUiWrapper.module.scss';
 
-export interface TCoreWrapperProps extends JSX.IntrinsicAttributes {
+export interface TCoreUiWrapperProps extends JSX.IntrinsicAttributes {
   className?: string;
   children: React.ReactNode;
   // themeMode?: TMuiThemeMode;
 }
 
-export function CoreWrapper(props: TCoreWrapperProps) {
+export function CoreUiWrapper(props: TCoreUiWrapperProps) {
   const { children, className } = props;
   // TODO: `useGlobalCssClasses`?
   return (
